Added very basic (but written by me) scripts
So added two controllers, for Enemies and Player (Kuro), and I'm struggling with the Raycast, as it doesn't land, however it prints it when I launch (play) the code in Unity Editor. And also the code for movement and jumping is really REALLY basic, so if there's any ideas to make it more fluent as it moves a bit laggy and takes a couple miliseconds to start moving when you press the movement keys.