Updated KuroController
Modified the EnemyDetected method in KuroController in order to set a Raycast, however, the Raycast is hitting the origin collider and ignoring all the other elements on the map. Upon adding a specific LayerMask, I will get a NullExceptionError saying that there's a reference to an object that has no instances. Stuck here for now.