Updated KuroController code
Added a function to test if raycast is detecting EnemyCollider. It does. 
Modified a function and renamed it to KuroSpotsEnemy(). Aiming to make it a boolean where a ray is cast regardless of where is Kuro looking at, so he's able to "see" enemies both in the X axis and -X axis when player turns around, and "react" with a short animation upon gazing enemies (once the enemy collider is detected).