Update KuroController
Edited and made code for method KuroSpotsEnemy() so that when player-shot raycast detects an enemy it would return true. Tried to check it with Debug.Logs but I was only able to see the boolean results there as the returns are not shown in the Console of Unity, but the right values show on Debug.Logs. 
Added a LayerMask variable doing a bitshift.