Updated KuroController()
Edited and made code for method KuroSpotsEnemy() so that when player-shot raycast detects an enemy it would return true. Tried to check it with Debug.Logs but I was only able to see the boolean results there as the returns are not shown in the Console of Unity, but the right values show on Debug.Logs. 
-Added a LayerMask variable doing a bitshift.
-The ray detects the collider of the Enemy BUT only after there has been a collider vs collider collision. 
- When it turns, it somehow detects the EnemyCollider in both cases (flip.X == true and flip.X == false). Is it because of the raycast being called to Mathf.Infinity in both cases?