Updated KuroController
Partially solved the physics bug issue. However, issue with the collision causing the player to rotate and then "fall" to the ground with its back or face on the floor persists. 
Player is now able to "stand up" after falling product from colliding with Enemy (this falling part should not happen, need to get deeper understanding on physics2D and 3D).